fic chisiamo in sito, impianti continua

+++ database/migrations/2020_07_03_092707_create_implants_table.php
@@ -41,7 +41,7 @@ class CreateImplantsTable extends Migration
             $table->string('meta_titolo')->nullable();
             $table->string('meta_descrizione')->nullable();
             $table->string('meta_keywords')->nullable();
-            $table->bigIncrements('views');
+            $table->integer('views');
 
 
             $table->enum('pubblicato',['on',NULL])->nullable();
